moved all pydantic models to models.py
